+++ robotbase/src/main/java/com/team2357/frc2023/commands/intake/AlertCommand.java
@@ -1,4 +1,4 @@
-package com.team2357.frc2023.commands;
+package com.team2357.frc2023.commands.intake;
 
 import edu.wpi.first.wpilibj2.command.CommandBase;
