--- robotbase/src/main/java/com/team2357/frc2023/commands/intake/ReverseIntakeCommand.java
@@ -1,20 +1,0 @@
-package com.team2357.frc2023.commands.intake;
-
-import com.team2357.frc2023.subsystems.IntakeRollerSubsystem;
-import com.team2357.lib.commands.CommandLoggerBase;
-
-public class IntakeReverseCommand extends CommandLoggerBase {
-    public IntakeReverseCommand() {
-        addRequirements(IntakeRollerSubsystem.getInstance());
-    }
-
-    @Override
-    public void execute() {
-        IntakeRollerSubsystem.getInstance().runIntake(true);
-    }
-
-    @Override
-    public void end(boolean interrupted) {
-        IntakeRollerSubsystem.getInstance().stopIntake();
-    }
-}
