+++ robotbase/src/main/java/com/team2357/frc2023/commands/scoring/AutoScoreLowCommand.java
@@ -2,11 +2,10 @@ package com.team2357.frc2023.commands.scoring;
 
 import com.team2357.frc2023.Constants;
 import com.team2357.frc2023.commands.intake.DeployIntakeCommand;
-import com.team2357.frc2023.commands.intake.ReverseIntakeCommand;
+import com.team2357.frc2023.commands.intake.IntakeReverseCommand;
 import com.team2357.frc2023.commands.scoring.util.ExtendArmToPositionCommand;
 import com.team2357.frc2023.commands.scoring.util.OpenClawCommand;
 
-import edu.wpi.first.wpilibj2.command.ParallelCommandGroup;
 import edu.wpi.first.wpilibj2.command.ParallelRaceGroup;
 import edu.wpi.first.wpilibj2.command.SequentialCommandGroup;
 import edu.wpi.first.wpilibj2.command.WaitCommand;
