+++ robotbase/src/main/java/com/team2357/frc2023/commands/scoring/AutoScoreMidCommand.java
@@ -1,7 +1,7 @@
 package com.team2357.frc2023.commands.scoring;
 
 import com.team2357.frc2023.Constants;
-import com.team2357.frc2023.commands.intake.ReverseIntakeCommand;
+import com.team2357.frc2023.commands.intake.IntakeReverseCommand;
 import com.team2357.frc2023.commands.scoring.util.ExtendArmToPositionCommand;
 import com.team2357.frc2023.commands.scoring.util.ExtendWristCommand;
 import com.team2357.frc2023.commands.scoring.util.OpenClawCommand;
