+++ robotbase/src/main/java/com/team2357/frc2023/controls/SwerveDriveControls.java
@@ -1,7 +1,7 @@
 package com.team2357.frc2023.controls;
 
-import com.team2357.frc2023.commands.IntakeReverseCommand;
-import com.team2357.frc2023.commands.IntakeRunCommand;
+import com.team2357.frc2023.commands.intake.IntakeReverseCommand;
+import com.team2357.frc2023.commands.intake.IntakeRunCommand;
 import com.team2357.frc2023.subsystems.IntakeRollerSubsystem;
 import com.team2357.frc2023.commands.scoring.teleopAutoScore.TeleopAutoScoreCommandGroup;
 import com.team2357.frc2023.subsystems.SwerveDriveSubsystem;
@@ -19,7 +19,6 @@ public class SwerveDriveControls {
     private JoystickButton m_backButton;
     private JoystickButton m_rightBumper;
     private JoystickButton m_leftBumper;
-    private JoystickButton m_rightTrigger;
 
     private AxisThresholdTrigger m_leftTrigger;
     private AxisThresholdTrigger m_rightTrigger;
@@ -35,7 +34,7 @@ public class SwerveDriveControls {
         m_rightBumper = new JoystickButton(m_controller, XboxRaw.BumperRight.value);
         m_leftBumper = new JoystickButton(m_controller, XboxRaw.BumperLeft.value);
 
-        m_rightTrigger = new JoystickButton(m_controller, XboxRaw.TriggerRight.value);
+        m_rightTrigger = new AxisThresholdTrigger(controller, Axis.kRightTrigger, 0.1);
 
         mapControls();
 
@@ -46,8 +45,8 @@ public class SwerveDriveControls {
         m_backButton.whileTrue(new InstantCommand(() -> SwerveDriveSubsystem.getInstance().zeroGyroscope()));
         
         // Intake commands
-        m_rightBumper.whileTrue(new RunIntakeCommand());
-        m_leftBumper.whileTrue(new ReverseIntakeCommand());
+        m_rightBumper.whileTrue(new IntakeRunCommand());
+        m_leftBumper.whileTrue(new IntakeReverseCommand());
 
         // Teleop auto
         m_rightTrigger.whileTrue(new TeleopAutoScoreCommandGroup());
